refactor(prompts): leave lib/prompts as the chat prompt and nothing else
The article analyse, summarise and report prompts, the web-search
question and answer prompts, the extract-content prompt and the two
drawing prompts all sat beside getSystemPrompt while backing routes that
have nothing to do with chat. They move to the services that own them,
which takes lib/prompts/index.ts from 262 lines to 62.

The layout heuristic goes with them. It compacted the metadata section
below a 12k context window and the principles section below 9k - the
mechanism ADR 0018 removed from artifact guidance, on the grounds that
the compact variant drops the detail that prevents a broken result,
still applied to the sections that tell the model how to behave.

The QwQ thinking-tag quirk keeps one implementation rather than two: the
streaming path now shares the predicate with the formatter instead of
repeating the model-name test inline. Deleting it outright needs a model
capability rather than a substring match, which is a catalogue change
this does not make.
