+++ apps/api/src/services/apps/articles/prompts.ts
@@ -1,0 +1,116 @@
+import {
+  buildAssistantMetadataSection,
+  type PromptModelMetadata,
+} from "~/lib/prompts/sections/metadata";
+import type { ChatMode } from "~/types";
+
+function buildArticlePromptMetadata(
+  metadata: PromptModelMetadata | undefined,
+  mode: ChatMode,
+): string {
+  return buildAssistantMetadataSection({
+    request: {
+      mode,
+      verbosity: "high",
+    },
+    modelId: metadata?.modelId,
+    modelConfig: metadata?.modelConfig,
+  });
+}
+
+export function analyseArticlePrompt(article: string, metadata?: PromptModelMetadata): string {
+  const metadataSection = buildArticlePromptMetadata(metadata, "normal");
+
+  return `<s> [INST] ${metadataSection}Your task is provide a comprehensive analysis that identifies any potential bias, political leanings, and the tone of the content, evaluating the presence of bias and political alignment in the article provided.
+
+Use the content provided under the heading "Article" and only that content to conduct your analysis. Do not embellish or add detail beyond the source material. The term "Article" is a placeholder for the actual content and should not be included in your output.
+
+Always assist with care, respect, and truth. Respond with utmost utility yet securely. Avoid harmful, unethical, prejudiced, or negative content. Ensure replies promote fairness and positivity.
+
+### Article ###:
+${article}
+
+### Instructions ###:
+1. Carefully read the "Article" and note any language, phrasing, or content that may indicate bias or political alignment.
+2. Do not include any conversational phrases, personal comments, or introductions. Only provide the summary and necessary sections as outlined below.
+3. Provide your response in English only.
+4. Your analysis must include the following sections:
+   - **Introduction**: Briefly introduce the "Article" and its main topic or focus.
+   - **Bias Detection**: Identify any signs of bias in the language, tone, or presentation of facts. This includes loaded language, unbalanced reporting, omission of key perspectives, or any use of subjective language that could sway the reader's opinion.
+   - **Political Alignment**: Analyze the content for indicators of political alignment. This can include the portrayal of political figures, policies, or ideologies in a favorable or unfavorable light, as well as any endorsement or criticism that aligns with specific political ideologies.
+   - **Examples and Evidence**: Provide specific examples from the text to support your findings. This should include direct quotes or paraphrased content that clearly illustrates the bias or political alignment identified.
+   - **Conclusion**: Summarize your findings, highlighting the overall bias and political alignment, if any, and the potential impact on the reader's perception.
+5. Format your analysis in clear, organized paragraphs with appropriate headings for each section. Use the Markdown format.
+6. Maintain a neutral and objective tone throughout your analysis. Avoid subjective judgments or interpretations that are not directly supported by evidence from the "Article".
+
+[/INST]
+
+Analysis: </s>`;
+}
+
+export function summariseArticlePrompt(article: string, metadata?: PromptModelMetadata): string {
+  const metadataSection = buildArticlePromptMetadata(metadata, "normal");
+
+  return `<s> [INST] ${metadataSection}Your task is to provide a professional summary of the article provided.
+
+Use the content provided under the heading "Article" and only that content to conduct your analysis. Do not embellish or add detail beyond the source material. The term "Article" is a placeholder for the actual content and should not be included in your output.
+
+Always assist with care, respect, and truth. Ensure replies are useful, secure, and promote fairness and positivity, avoiding harmful, unethical, prejudiced, or negative content.
+
+### Article ###:
+${article}
+
+### Instructions ###:
+1. Read the "Article" carefully, noting the main topics and subjects.
+2. Do not include any conversational phrases, personal comments, or introductions. Only provide the summary and necessary sections as outlined below.
+3. Provide your response in English only.
+4. Your summary must be between 300-400 words long (excluding keywords). This range is for internal use and should not be mentioned in the output.
+4a. Ensure all points are complete within this limit, even if fewer points are included or slight extensions are made. Do not cut points short, and do not include the word count in the output.
+5. Your report must include the following sections:
+   - **Introduction**: Briefly introduce the "Article" in no more than 30 words.
+   - **Key Findings**: Summarize 3-5 key findings and insights concisely, using approximately 220 words. Prioritize the most impactful points.
+   - **Quotes**: Include at least one brief illustrative quote that significantly enhances a key finding. Paraphrase where possible to maintain brevity.
+   - **Context and Inferences**: Provide any relevant context or inferences in about 30 words, if space allows.
+   - **Keywords**: List relevant subject keywords for content tagging, focusing on core topics and themes.
+6. Format your summary in clear paragraphs with headings for each section. Use Markdown format. Bullet points may be used for clarity where appropriate.
+
+[/INST]
+
+Summary: </s>`;
+}
+
+export function generateArticleReportPrompt(
+  articles: string,
+  metadata?: PromptModelMetadata,
+): string {
+  const metadataSection = buildArticlePromptMetadata(metadata, "normal");
+
+  return `<s> [INST] ${metadataSection}You have been given an article to summarize as a professional researcher. Your task is to provide a report that summarises a collection of article summaries.
+
+Use the content provided under the heading "Article Summaries" and only that content to conduct your analysis. Do not embellish or add detail beyond the source material. The term "Article Summaries" is a placeholder for the actual content and should not be included in your output.
+
+When encountering a "REPLACED_IMAGE" tag, disregard the tag but consider the text below it as a description of the removed image.
+
+Always assist with care, respect, and truth. Ensure replies are useful, secure, and promote fairness and positivity, avoiding harmful, unethical, prejudiced, or negative content.
+
+### Article Summaries ###:
+${articles}
+
+### Instructions ###:
+1. Read the "Article Summaries" carefully, focusing on the main topics and subjects.
+2. Your summary must be between 800-1500 words long (excluding keywords). This range is for internal use and should not be mentioned in the output.
+2a. Ensure all points are complete within this limit, even if fewer points are included or slight extensions are made. Do not cut points short.
+3. Your report must include the following sections:
+   - **Introduction**: Briefly introduce the "Article Summaries" in no more than 100 words.
+   - **Key Findings**: Summarize 3-5 key findings and insights concisely, using approximately 500-600 words. Prioritize the most impactful points.
+   - **Quotes**: Include at least one brief illustrative quote that significantly enhances a key finding. Paraphrase where possible to maintain brevity.
+   - **Context and Inferences**: Provide any relevant context or inferences in about 100-200 words, if space allows.
+   - **Keywords**: List relevant subject keywords for content tagging, focusing on core topics and themes.
+
+4. Format your summary in clear paragraphs with headings for each section. Use Markdown format. Bullet points may be used for clarity where appropriate.
+5. Regularly monitor the word count to ensure the final summary adheres to the word limit, without including the count in the output.
+
+[/INST]
+
+Summary: </s>`;
+}
