+++ apps/api/src/services/apps/drawing/prompts.ts
@@ -1,0 +1,31 @@
+export function drawingDescriptionPrompt(): string {
+  return `You are an advanced image analysis AI capable of providing accurate and concise descriptions of visual content. Your task is to describe the given image in a single, informative sentence.
+
+Instructions:
+1. Carefully analyze the image content.
+2. Identify key elements, shapes, objects, or patterns present in the image.
+3. Pay special attention to distinguishable features, even if the image appears mostly dark or monochromatic.
+4. Formulate a single sentence that accurately describes the main elements of the image.
+
+Your final output should be a single sentence describing the image.
+
+Example output structure:
+
+[A single sentence describing the main elements of the image]`;
+}
+
+export function guessDrawingPrompt(usedGuesses: Set<string>): string {
+  return `You will be provided with a description of an image. Your task is to guess what the image depicts using only one word. Follow these steps:
+
+1. Carefully review the image provided.
+
+2. Based on the image, think about the most likely object, animal, place, food, activity, or concept that the image represents.
+
+3. Choose a single word that best describes or identifies the main subject of the image.
+
+4. Provide your guess as a single word response. Do not include any explanations, punctuation, or additional text.
+
+IMPORTANT: Do not use any of these previously guessed words: ${Array.from(usedGuesses).join(", ")}
+
+Your response should contain only one word, which represents your best guess for the image described. Ensure that your answer is concise and accurately reflects the main subject of the image.`;
+}
