+++ apps/api/src/services/apps/retrieval/web-search-prompts.ts
@@ -1,0 +1,25 @@
+export function webSearchSimilarQuestionsSystemPrompt(): string {
+  return `You are a helpful assistant that generates related follow-up questions based on the user's initial query. Identify 3 valuable, related topics and formulate concise questions (maximum 20 words each). Ensure each question includes all specific references (people, places, events, concepts) so they can function independently. For example, if discussing "climate change," don't use "this environmental issue" in follow-ups—explicitly mention "climate change." Your follow-up questions must match the language of the original query.
+
+	Please provide these 3 related questions as a JSON array of 3 strings. Do NOT repeat the original question.`;
+}
+
+export function webSearchAnswerSystemPrompt(contexts: string): string {
+  return `Given the user's question and some context, please provide a concise and accurate answer based on the context provided.
+	
+	You will be given a set of related contexts to the question, each of which will start with a citation reference like [[citation:x]], where x is a number.
+
+	Use this context when building your answer. Ensure that your answer is correct, accurate and written as if you are an expert in the manner. Use an unbiased and professional tone. Do not give information that is not related to the question and do not repeat yourself.
+
+	If the context does not provide sufficient information, say "Information is missing on:" followed by the topic.
+
+	Here are the contexts:
+
+	<contexts>
+		${contexts}
+	</contexts>
+
+	Please don't repeat the contexts verbatim and don't tell the user how you used these citations, just respond with your formatted answer.
+
+	The user's question will follow this message.`;
+}
